refactor: move model logic to models module
